speakup-services/requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.1
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.1
packaging==24.1
pandas==2.2.2
praat-parselmouth==0.4.4
python-dateutil==2.9.0.post0
pytz==2024.1
scipy==1.14.0
six==1.16.0
tzdata==2024.1
watchdog==4.0.1
Werkzeug==3.0.3
